requirements.txt
Pillow==7.0.0
pyserial==3.4
thermalprinter==0.2.0
